fixed some errors in tester

+++ libft/ft_max.c
@@ -3,14 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   ft_max.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mel-idri <mel-idri@student.42.fr>          +#+  +:+       +#+        */
+/*   By: ylagtab <ylagtab@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/22 21:22:23 by ylagtab           #+#    #+#             */
-/*   Updated: 2019/11/20 23:20:26 by mel-idri         ###   ########.fr       */
+/*   Updated: 2019/11/21 22:15:13 by ylagtab          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_max(int nbr1, int nbr2)
+int	ft_max(unsigned long nbr1, unsigned long nbr2)
 {
 	return (nbr1 > nbr2 ? nbr1 : nbr2);
 }
